requirements.txt
PyQt6>=6.4.0
Pillow>=9.0.0
numpy>=1.21.0
av>=11.0.0
pyinstaller>=6.0.0
